backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
distro==1.9.0
Faker==37.1.0
Flask==3.1.0
flask-cors==5.0.1
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
MarkupSafe==3.0.2
numpy==2.2.4
openai==1.75.0
pandas==2.2.3
pydantic==2.11.3
pydantic_core==2.33.1
pyspark==3.5.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
Werkzeug==3.1.3
